rework

+++ install.sh
@@ -1,7 +1,10 @@
-#!/usr/bin/sh
+#!/usr/bin/bash
+
+set -e
+source config.sh
 
 echo 'loading keymap'
-loadkeys de
+loadkeys $keymap
 
 echo 'enabling ntp'
 timedatectl set-ntp true
@@ -11,6 +14,19 @@ read -n 1 -srp "Press any key to continue"
 cfdisk
 
 read -p "Enter the partition you want to install to: " partition
+
+if [ "$encrypt_root" = true ]; then
+    echo 'Please read into the arch wiki page on drive preperation for encryption'
+    read -n 1 -srp "Press any key to continue"
+    cryptsetup luksFormat $partition
+    cryptsetup open $partition cryptlvm
+    pvcreate /dev/mapper/cryptlvm
+    vgcreate vg_root /dev/mapper/cryptlvm
+    lvcreate -l 100%FREE vg_root -n root
+
+    partition="/dev/vg_root/root" # from here on $partition is the lvm name
+fi
+
 echo 'creating file system'
 mkfs.ext4 $partition
 
@@ -24,7 +40,7 @@ mount $partition /mnt
 echo 'mounting /boot'
 mkdir /mnt/boot
 mount $boot_partition
-            
+
 echo 'downloading packages'
 pacstrap /mnt base
